+++ package.json
@@ -9,7 +9,6 @@
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.6.0",
-    "axios": "^0.21.1",
     "luxon": "^1.25.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
