+++ src/components/DatosUsuario.js
@@ -1,4 +1,6 @@
 import { Link, useParams } from 'react-router-dom';
+import { useRecoilValue } from 'recoil';
+import { usuarioPorId } from '../state/usuarios';
 import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import CardActionArea from '@material-ui/core/CardActionArea';
@@ -6,11 +8,6 @@ import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import Typography from '@material-ui/core/Typography';
 import { Button } from '@material-ui/core';
-import { useEffect, useState } from 'react';
-import { Alert } from '@material-ui/lab';
-import { getUsuarioPorId } from '../services/usuarios';
-import { getUsuarioPorId as getUsuarioPorId_fake } from '../services/usuarios-fake';
-import { getDataFromBackend, usuariosFijos } from '../constants/constants';
 
 const useStyles = makeStyles({
   root: {
@@ -21,28 +18,11 @@ const useStyles = makeStyles({
 export default function DatosUsuario() {
   const { id } = useParams();
   const classes = useStyles();
+  const usuario = useRecoilValue(usuarioPorId(id));
 
-  const [usuario, setUsuario] = useState(null);
-  const [hasError, setHasError] = useState(false);
-
-  useEffect(() => {
-    async function fetchUsuario() {
-      try {
-        const getFunction = getDataFromBackend
-          ? getUsuarioPorId
-          : getUsuarioPorId_fake;
-        const usuario = await getFunction(id);
-        setUsuario(usuario);
-      } catch (err) {
-        setHasError(true);
-      }
-    }
-    fetchUsuario();
-  }, [id]);
-
-  const usuarioRendering = () => {
-    return [
-      <Card className={classes.root} key="datosUsuario">
+  return (
+    <>
+      <Card className={classes.root}>
         <CardActionArea>
           <CardMedia
             component="img"
@@ -61,31 +41,10 @@ export default function DatosUsuario() {
             </Typography>
           </CardContent>
         </CardActionArea>
-      </Card>,
-      <Button color="primary" component={Link} to="/" key="botonVolver">
+      </Card>
+      <Button color="primary" component={Link} to="/">
         Volver
-      </Button>,
-    ];
-  };
-
-  const errorRendering = () => {
-    return (
-      <Alert severity="warning">
-        No pudimos cargar el usuario. ¿Levantaste la API?{' '}
-        <span role="img" aria-label="thinking">
-          🤔
-        </span>
-      </Alert>
-    );
-  };
-
-  const loadingRendering = () => {
-    return <Alert severity="info">Cargando usuarie ...</Alert>;
-  };
-
-  return hasError
-    ? errorRendering()
-    : usuario == null
-    ? loadingRendering()
-    : usuarioRendering();
+      </Button>
+    </>
+  );
 }
