+++ src/components/ListadoUsuarios.js
@@ -2,7 +2,8 @@ import { makeStyles } from '@material-ui/core/styles';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import ListItemText from '@material-ui/core/ListItemText';
-
+import { useRecoilValue } from 'recoil';
+import { todosLosUsuarios } from '../state/usuarios';
 import {
   Avatar,
   Divider,
@@ -15,11 +16,6 @@ import { Alert } from '@material-ui/lab';
 import { DateTime } from 'luxon';
 import { Face } from '@material-ui/icons';
 import { Link } from 'react-router-dom';
-import { useEffect, useState } from 'react';
-
-import { getTodosLosUsuarios } from '../services/usuarios';
-import { getTodosLosUsuarios as getTodosLosUsuarios_fake } from '../services/usuarios-fake';
-import { getDataFromBackend } from '../constants/constants';
 
 const useStyles = makeStyles(() => ({
   root: {
@@ -41,32 +37,14 @@ function fechaFormatoHumano(fecha) {
 
 export default function ListadoUsuarios() {
   const classes = useStyles();
-  const [usuarios, setUsuarios] = useState(null);
-  const [hasError, setHasError] = useState(false);
+  const usuarios = useRecoilValue(todosLosUsuarios);
 
-  useEffect(() => {
-    async function fetchUsuarios() {
-      const getFunction = getDataFromBackend
-        ? getTodosLosUsuarios
-        : getTodosLosUsuarios_fake;
-      try {
-        const usuarios = await getFunction();
-        setUsuarios(usuarios);
-      } catch (err) {
-        setHasError(true);
-      }
-    }
-    fetchUsuarios();
-  }, []);
-
-  const usuariosRendering = () => {
-    return [
-      <Alert severity="info" className={classes.alert} key="alert">
-        {getDataFromBackend
-          ? 'Los usuarios que están más abajo vienen de la API.'
-          : 'Estos usuarios son fijos'}
-      </Alert>,
-      <List className={classes.root} key="usuarios">
+  return (
+    <Grid container>
+      <Alert severity="info" className={classes.alert}>
+        Los usuarios que están más abajo vienen de la API.
+      </Alert>
+      <List className={classes.root}>
         {usuarios.map((it, index) => (
           <div key={it.id}>
             <ListItem alignItems="flex-start">
@@ -87,9 +65,6 @@ export default function ListadoUsuarios() {
                 >
                   <Face />
                 </IconButton>
-                <Link style={{ marginLeft: '10px' }} to={`/usuarios/${it.id}`}>
-                  Ver detalle
-                </Link>
               </ListItemSecondaryAction>
             </ListItem>
             {/* Hack para que no muestre el divider en el último elemento */}
@@ -98,32 +73,7 @@ export default function ListadoUsuarios() {
             )}
           </div>
         ))}
-      </List>,
-    ];
-  };
-
-  const errorRendering = () => {
-    return (
-      <Alert severity="warning">
-        No pudimos cargar los usuarios. ¿Levantaste la API?{' '}
-        <span role="img" aria-label="thinking">
-          🤔
-        </span>
-      </Alert>
-    );
-  };
-
-  const loadingRendering = () => {
-    return <Alert severity="info">Cargando usuaries ...</Alert>;
-  };
-
-  return (
-    <Grid container>
-      {hasError
-        ? errorRendering()
-        : usuarios == null
-        ? loadingRendering()
-        : usuariosRendering()}
+      </List>
     </Grid>
   );
 }
