--- src/constants/constants.js
@@ -1,34 +1,0 @@
-export const getDataFromBackend = false;
-
-export const usuariosFijos = [
-  {
-    id: 1,
-    nombre: 'Juana',
-    apellido: 'Molina',
-    fechaNacimiento: '1971-04-21',
-    edad: 50,
-  },
-  {
-    id: 2,
-    nombre: 'China',
-    apellido: 'Zorrilla',
-    fechaNacimiento: '1929-04-09',
-    edad: 92,
-  },
-  {
-    id: 3,
-    nombre: 'Nelly',
-    apellido: 'Omar',
-    fechaNacimiento: '1921-11-17',
-    edad: 100,
-  },
-  {
-    id: 4,
-    nombre: 'Erica',
-    apellido: 'Rivas',
-    fechaNacimiento: '1979-10-29',
-    avatarUrl:
-      'https://static.wikia.nocookie.net/logopedia/images/c/c9/RIVERNORMAL.png',
-    edad: 42,
-  },
-];
