--- src/services/utils.js
@@ -1,8 +1,0 @@
-import axios from 'axios';
-
-const apiUrl = process.env.REACT_APP_API_URL;
-
-export async function getJsonFromApi(path) {
-  const response = await axios.get(`${apiUrl}/${path}`);
-  return response.data;
-}
